updates paperclip settings

+++ dbProcs/filteredcontentquery.sql
@@ -1,6 +1,6 @@
 ﻿
-DROP FUNCTION filteredcontentquery(mytoken text, 
-						lat double precision, lon double precision, radius double precision, 
+DROP FUNCTION filteredcontentquery(mytoken text,
+						lat double precision, lon double precision, radius double precision,
 						minalt double precision, maxalt double precision,
 						mintime timestamp, maxtime timestamp,
 						taglist text,
@@ -16,7 +16,7 @@ raw json call:
 http://54.221.254.230/images/filteredcontent.json?lat=43.32485&long=-79.81303&radius=500.0&altmin=-1000.0&altmax=1000.0&timemin=NULL&timemax=NULL&taglist=''&userlist=''&catlist=''&maxcount=10
 */
 CREATE OR REPLACE FUNCTION filteredcontentquery(mytoken text,
-						lat double precision, lon double precision, radius double precision, 
+						lat double precision, lon double precision, radius double precision,
 						minalt double precision, maxalt double precision,
 						mintime timestamp, maxtime timestamp,
 						taglist text,
@@ -25,17 +25,17 @@ CREATE OR REPLACE FUNCTION filteredcontentquery(mytoken text,
 						followingpics integer,
 						maxcount integer
 						)
-RETURNS TABLE(id bigint, content_type integer, latitude double precision, longitude double precision, altitude double precision)
+RETURNS TABLE(id int, content_type integer, latitude double precision, longitude double precision, altitude double precision)
 AS $$
 DECLARE
 	tagset text[];
 	tagarraylen integer;
-	
+
 	userset integer[];
 	userarraylen integer;
 
 	my_id integer;
-	
+
 	skiploc boolean;
 	skipsocial boolean;
 
@@ -45,29 +45,29 @@ BEGIN
 
 	skipsocial = NOT ((mypics = 1) OR (followingpics = 1));
 
-	SELECT u.id INTO my_id 
-	FROM users AS u 
+	SELECT u.id INTO my_id
+	FROM users AS u
 	WHERE authentication_token = mytoken;
 
 	tagset = string_to_array(trim(both ' ' from taglist), ' ');
 	tagarraylen = array_length(tagset, 1);
-	
+
 	userset = string_to_array(trim(both ' ' from userlist), ' ');
 	userarraylen = array_length(userset, 1);
-	
+
 	IF (mintime IS NULL) THEN
 		mintime = '-infinity'::timestamp;
 	END IF;
-		
+
 	IF (maxtime IS NULL) THEN
 		maxtime = 'infinity'::timestamp;
 	END IF;
-		
+
 	CREATE TEMP TABLE imagesinbox
 	ON COMMIT DROP
 	AS
-	(	
-		SELECT	i.id, 
+	(
+		SELECT	i.id,
 			i.privacy,
 			i.user_id
 		FROM	(SELECT * FROM buildboundingbox(lat, lon, radius) FETCH FIRST 1 ROWS ONLY) as bb,
@@ -81,9 +81,9 @@ BEGIN
 
 	CREATE TEMP TABLE imageset
 	ON COMMIT DROP
-	AS 
+	AS
 	(
-		SELECT	DISTINCT i.* 
+		SELECT	DISTINCT i.*
 		FROM	imagesinbox i
 			INNER JOIN users u ON (i.user_id = u.id)
 			LEFT OUTER JOIN connections c ON ((c.user_id = my_id) AND (c.connections_id = u.id))
@@ -98,7 +98,7 @@ BEGIN
 	);
 
 RETURN QUERY
-	SELECT	DISTINCT i.id AS id, 1::integer AS content_type, 
+	SELECT	DISTINCT i.id AS id, 1::integer AS content_type,
 			i.best_latitude AS latitude, i.best_longitude AS longitude, i.best_altitude AS altitude
 	FROM	imageset ims
 		INNER JOIN images i ON (ims.id = i.id)
